+++ docs/generators/typescript-inversify.md
@@ -13,10 +13,10 @@ sidebar_label: typescript-inversify
 |prependFormOrBodyParameters|Add form or body parameters to the beginning of the parameter list.| |false|
 |modelPropertyNaming|Naming convention for the property: 'camelCase', 'PascalCase', 'snake_case' and 'original', which keeps the original name| |camelCase|
 |supportsES6|Generate code that conforms to ES6.| |false|
-|npmName|The name under which you want to publish generated npm package| |null|
-|npmVersion|The version of your npm package| |null|
+|npmName|The name under which you want to publish generated npm package. Required to generate a full package| |null|
+|npmVersion|The version of your npm package. If not provided, using the version from the OpenAPI specification file.| |1.0.0|
+|snapshot|When setting this property to true, the version will be suffixed with -SNAPSHOT.yyyyMMddHHmm| |false|
 |npmRepository|Use this property to set an url your private npmRepo in the package.json| |null|
-|snapshot|When setting this property to true the version will be suffixed with -SNAPSHOT.yyyyMMddHHmm| |false|
 |withInterfaces|Setting this property to true will generate interfaces next to the default class implementations.| |false|
 |usePromise|Setting this property to use promise instead of observable inside every service.| |false|
 |taggedUnions|Use discriminators to create tagged unions instead of extending interfaces.| |false|
