+++ modules/openapi-generator/src/test/java/org/openapitools/codegen/typescript/fetch/TypeScriptFetchClientCodegenTest.java
@@ -1,5 +1,7 @@
 package org.openapitools.codegen.typescript.fetch;
 
+import io.swagger.v3.oas.models.OpenAPI;
+import org.openapitools.codegen.TestUtils;
 import org.openapitools.codegen.languages.TypeScriptFetchClientCodegen;
 import org.testng.Assert;
 import org.testng.annotations.Test;
@@ -8,11 +10,14 @@ import org.testng.annotations.Test;
 public class TypeScriptFetchClientCodegenTest {
     @Test
     public void testSnapshotVersion() {
+        OpenAPI api = TestUtils.createOpenAPI();
         TypeScriptFetchClientCodegen codegen = new TypeScriptFetchClientCodegen();
+
         codegen.additionalProperties().put("npmName", "@openapi/typescript-fetch-petstore");
         codegen.additionalProperties().put("snapshot", true);
         codegen.additionalProperties().put("npmVersion", "1.0.0-SNAPSHOT");
         codegen.processOpts();
+        codegen.preprocessOpenAPI(api);
 
         Assert.assertTrue(codegen.getNpmVersion().matches("^1.0.0-SNAPSHOT.[0-9]{12}$"));
 
@@ -21,6 +26,7 @@ public class TypeScriptFetchClientCodegenTest {
         codegen.additionalProperties().put("snapshot", true);
         codegen.additionalProperties().put("npmVersion", "3.0.0-M1");
         codegen.processOpts();
+        codegen.preprocessOpenAPI(api);
 
         Assert.assertTrue(codegen.getNpmVersion().matches("^3.0.0-M1-SNAPSHOT.[0-9]{12}$"));
 
@@ -28,11 +34,14 @@ public class TypeScriptFetchClientCodegenTest {
 
     @Test
     public void testWithoutSnapshotVersion() {
+        OpenAPI api = TestUtils.createOpenAPI();
         TypeScriptFetchClientCodegen codegen = new TypeScriptFetchClientCodegen();
+
         codegen.additionalProperties().put("npmName", "@openapi/typescript-fetch-petstore");
         codegen.additionalProperties().put("snapshot", false);
         codegen.additionalProperties().put("npmVersion", "1.0.0-SNAPSHOT");
         codegen.processOpts();
+        codegen.preprocessOpenAPI(api);
 
         Assert.assertTrue(codegen.getNpmVersion().matches("^1.0.0-SNAPSHOT$"));
 
@@ -41,6 +50,7 @@ public class TypeScriptFetchClientCodegenTest {
         codegen.additionalProperties().put("snapshot", false);
         codegen.additionalProperties().put("npmVersion", "3.0.0-M1");
         codegen.processOpts();
+        codegen.preprocessOpenAPI(api);
 
         Assert.assertTrue(codegen.getNpmVersion().matches("^3.0.0-M1$"));
