+++ modules/openapi-generator/src/test/java/org/openapitools/codegen/typescript/typescriptnode/TypeScriptNodeClientCodegenTest.java
@@ -1,5 +1,7 @@
 package org.openapitools.codegen.typescript.typescriptnode;
 
+import io.swagger.v3.oas.models.OpenAPI;
+import org.openapitools.codegen.TestUtils;
 import org.openapitools.codegen.languages.TypeScriptNodeClientCodegen;
 import org.testng.Assert;
 import org.testng.annotations.Test;
@@ -20,11 +22,14 @@ public class TypeScriptNodeClientCodegenTest {
 
    @Test
    public void testSnapshotVersion() {
+      OpenAPI api = TestUtils.createOpenAPI();
       TypeScriptNodeClientCodegen codegen = new TypeScriptNodeClientCodegen();
+
       codegen.additionalProperties().put("npmName", "@openapi/typescript-angular-petstore");
       codegen.additionalProperties().put("snapshot", true);
       codegen.additionalProperties().put("npmVersion", "1.0.0-SNAPSHOT");
       codegen.processOpts();
+      codegen.preprocessOpenAPI(api);
 
       Assert.assertTrue(codegen.getNpmVersion().matches("^1.0.0-SNAPSHOT.[0-9]{12}$"));
 
@@ -33,6 +38,7 @@ public class TypeScriptNodeClientCodegenTest {
       codegen.additionalProperties().put("snapshot", true);
       codegen.additionalProperties().put("npmVersion", "3.0.0-M1");
       codegen.processOpts();
+      codegen.preprocessOpenAPI(api);
 
       Assert.assertTrue(codegen.getNpmVersion().matches("^3.0.0-M1-SNAPSHOT.[0-9]{12}$"));
 
@@ -40,11 +46,14 @@ public class TypeScriptNodeClientCodegenTest {
 
    @Test
    public void testWithoutSnapshotVersion() {
+      OpenAPI api = TestUtils.createOpenAPI();
       TypeScriptNodeClientCodegen codegen = new TypeScriptNodeClientCodegen();
+
       codegen.additionalProperties().put("npmName", "@openapi/typescript-angular-petstore");
       codegen.additionalProperties().put("snapshot", false);
       codegen.additionalProperties().put("npmVersion", "1.0.0-SNAPSHOT");
       codegen.processOpts();
+      codegen.preprocessOpenAPI(api);
 
       Assert.assertTrue(codegen.getNpmVersion().matches("^1.0.0-SNAPSHOT$"));
 
@@ -53,6 +62,7 @@ public class TypeScriptNodeClientCodegenTest {
       codegen.additionalProperties().put("snapshot", false);
       codegen.additionalProperties().put("npmVersion", "3.0.0-M1");
       codegen.processOpts();
+      codegen.preprocessOpenAPI(api);
 
       Assert.assertTrue(codegen.getNpmVersion().matches("^3.0.0-M1$"));
